requirements.txt
streamlit
langchain
openai
python-dotenv
ipykernel
langchain-openai
langchain-google-genai
pillow
youtube-transcript-api
langchain-community
faiss-cpu
tiktoken
pydantic
fastapi
uvicorn
chromadb
unstructured
